+++ android/src/main/java/com/cloudwebrtc/webrtc/MethodCallHandlerImpl.java
@@ -14,6 +14,8 @@ import android.media.MediaRecorder;
 import android.media.AudioAttributes;
 import android.media.AudioDeviceInfo;
 import android.os.Build;
+import android.os.Handler;
+import android.os.Looper;
 import android.util.Log;
 import android.util.LongSparseArray;
 import android.view.Surface;
@@ -51,6 +53,8 @@ import org.webrtc.DtmfSender;
 import org.webrtc.EglBase;
 import org.webrtc.IceCandidate;
 import org.webrtc.Logging;
+import org.webrtc.Logging.Severity;
+import org.webrtc.Loggable;
 import org.webrtc.MediaConstraints;
 import org.webrtc.MediaConstraints.KeyValuePair;
 import org.webrtc.MediaStream;
@@ -90,6 +94,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.UUID;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.EventChannel;
@@ -124,9 +130,11 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
   private FlutterRTCVideoPipe videoPipe;
   private CameraUtils cameraUtils;
 
-  private AudioDeviceModule audioDeviceModule;
+  private JavaAudioDeviceModule audioDeviceModule;
 
   private FlutterRTCFrameCryptor frameCryptor;
+
+  private FlutterDataPacketCryptor dataPacketCryptor;
 
   private Activity activity;
 
@@ -135,6 +143,21 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
   private CustomVideoDecoderFactory videoDecoderFactory;
 
   public AudioProcessingController audioProcessingController;
+
+  public static class LogSink implements Loggable {
+    @Override
+    public void onLogMessage(String message, Severity sev, String tag) {
+      ConstraintsMap params = new ConstraintsMap();
+      params.putString("event", "onLogData");
+      params.putString("data", message);
+      FlutterWebRTCPlugin.sharedSingleton.sendEvent(params.toMap());
+    }
+  }
+
+  ExecutorService executor = Executors.newSingleThreadExecutor();
+  Handler mainHandler = new Handler(Looper.getMainLooper());
+
+  public static LogSink logSink = new LogSink();
 
   MethodCallHandlerImpl(Context context, BinaryMessenger messenger, TextureRegistry textureRegistry) {
     this.context = context;
@@ -164,7 +187,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     mPeerConnectionObservers.clear();
   }
   private void initialize(boolean bypassVoiceProcessing, int networkIgnoreMask, boolean forceSWCodec, List<String> forceSWCodecList,
-  @Nullable ConstraintsMap androidAudioConfiguration) {
+  @Nullable ConstraintsMap androidAudioConfiguration, Severity logSeverity) {
     if (mFactory != null) {
       return;
     }
@@ -172,7 +195,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     PeerConnectionFactory.initialize(
             InitializationOptions.builder(context)
                     .setEnableInternalTracer(true)
-                    .setFieldTrials("WebRTC-Fec-03-Advertised/Enabled/") // Enable FEC
+                    .setInjectableLogger(logSink, logSeverity)
                     .createInitializationOptions());
 
     videoPipe = new FlutterRTCVideoPipe();
@@ -181,6 +204,8 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     cameraUtils = new CameraUtils(getUserMediaImpl, activity);
 
     frameCryptor = new FlutterRTCFrameCryptor(this);
+
+    dataPacketCryptor = new FlutterDataPacketCryptor(frameCryptor);
 
     AudioAttributes audioAttributes = null;
     if (androidAudioConfiguration != null) {
@@ -346,7 +371,15 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
         if(options.get("bypassVoiceProcessing") != null) {
           enableBypassVoiceProcessing = (boolean)options.get("bypassVoiceProcessing");
         }
-        initialize(enableBypassVoiceProcessing, networkIgnoreMask, forceSWCodec, forceSWCodecList, androidAudioConfiguration);
+
+        Severity logSeverity = Severity.LS_NONE;
+        if (constraintsMap.hasKey("logSeverity")
+                && constraintsMap.getType("logSeverity") == ObjectType.String) {
+          String logSeverityStr = constraintsMap.getString("logSeverity");
+          logSeverity = str2LogSeverity(logSeverityStr);
+        }
+
+        initialize(enableBypassVoiceProcessing, networkIgnoreMask, forceSWCodec, forceSWCodecList, androidAudioConfiguration, logSeverity);
         result.success(null);
         break;
       }
@@ -836,8 +869,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
       case "stopRecordToFile":
         Integer recorderId = call.argument("recorderId");
         String albumName = call.argument("albumName");
-        getUserMediaImpl.stopRecording(recorderId, albumName);
-        result.success(null);
+        getUserMediaImpl.stopRecording(recorderId, albumName, () -> result.success(null));
         break;
       case "captureFrame": {
         String path = call.argument("path");
@@ -1080,8 +1112,33 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
         }
         break;
       }
+      case "startLocalRecording": {
+        executor.execute(() -> {
+          audioDeviceModule.prewarmRecording();
+          mainHandler.post(() -> {
+            result.success(null);
+          });
+        });
+        break;
+      }
+      case "stopLocalRecording": {
+        executor.execute(() -> {
+          audioDeviceModule.requestStopRecording();
+          mainHandler.post(() -> {
+            result.success(null);
+          });
+        });
+        break;
+      }
+      case "setLogSeverity": {
+        //now it's possible to setup logSeverity only via PeerConnectionFactory.initialize method
+        //Log.d(TAG, "no implementation for 'setLogSeverity'");
+        break;
+      }
       default:
         if(frameCryptor.handleMethodCall(call, result)) {
+          break;
+        } else if(dataPacketCryptor.handleMethodCall(call, result)) {
           break;
         }
         result.notImplemented();
@@ -1636,7 +1693,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
       ConstraintsMap audioOutputMap = new ConstraintsMap();
       audioOutputMap.putString("label", audioOutput.getName());
       audioOutputMap.putString("deviceId", AudioDeviceKind.fromAudioDevice(audioOutput).typeName);
-      audioOutputMap.putString("groupId", "" + AudioDeviceKind.fromAudioDevice(audioOutput).typeName);
+      audioOutputMap.putString("groupId", AudioDeviceKind.fromAudioDevice(audioOutput).typeName);
       audioOutputMap.putString("kind", "audiooutput");
       array.pushMap(audioOutputMap);
     }
@@ -2087,6 +2144,22 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
       if (renderer.checkVideoTrack(trackId, "local")) {
         renderer.setStream(null, null);
       }
+    }
+  }
+
+  private Severity str2LogSeverity(String severity) {
+    switch (severity) {
+      case "verbose":
+        return Severity.LS_VERBOSE;
+      case "info":
+        return Severity.LS_INFO;
+      case "warning":
+        return Severity.LS_WARNING;
+      case "error":
+        return Severity.LS_ERROR;
+      case "none":
+      default:
+        return Severity.LS_NONE;
     }
   }