+++ common/cpp/src/flutter_webrtc.cc
@@ -1,8 +1,13 @@
 #include "flutter_webrtc.h"
+#include "flutter_data_channel.h"
 
 #include "flutter_webrtc_plus/flutter_web_r_t_c_plugin.h"
 
 namespace flutter_webrtc_plus_plugin {
+
+static EventChannelProxy* eventChannelProxy = nullptr;
+
+static EventChannelProxy* eventChannelProxy = nullptr;
 
 FlutterWebRTC::FlutterWebRTC(FlutterWebRTCPlugin* plugin)
     : FlutterWebRTCBase::FlutterWebRTCBase(plugin->messenger(),
@@ -24,6 +29,11 @@ void FlutterWebRTC::HandleMethodCall(
     const EncodableMap params =
         GetValue<EncodableMap>(*method_call.arguments());
     const EncodableMap options = findMap(params, "options");
+    std::string severityStr = findString(options, "logSeverity");
+    if (severityStr.empty() == false) {
+      RTCLoggingSeverity severity = str2LogSeverity(severityStr);
+      initLoggerCallback(severity);
+    }
     result->Success();
   } else if (method_call.method_name().compare("createPeerConnection") == 0) {
     if (!method_call.arguments()) {
@@ -352,7 +362,8 @@ void FlutterWebRTC::HandleMethodCall(
       return;
     }
     DataChannelSend(data_channel, type, data, std::move(result));
-  } else if (method_call.method_name().compare("dataChannelGetBufferedAmount") == 0) {
+  } else if (method_call.method_name().compare(
+                 "dataChannelGetBufferedAmount") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
       return;
@@ -1265,6 +1276,18 @@ void FlutterWebRTC::HandleMethodCall(
     state[EncodableValue("state")] =
         peerConnectionStateString(pc->peer_connection_state());
     result->Success(EncodableValue(state));
+  } else if (method_call.method_name().compare("setLogSeverity") == 0) {
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Bad arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+    std::string severityStr = findString(params, "severity");
+    if (severityStr.empty() == false) {
+      RTCLoggingSeverity severity = str2LogSeverity(severityStr);
+      initLoggerCallback(severity);
+    }
   } else if (method_call.method_name().compare("setThinValue") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
@@ -1275,7 +1298,7 @@ void FlutterWebRTC::HandleMethodCall(
     const double value = findDouble(params, "value");
     SetThinFaceValue(value);
     result->Success();
-  } else if (method_call.method_name().compare("setBigEyeValue") == 0) {
+  }  else if (method_call.method_name().compare("setBigEyeValue") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
       return;
@@ -1332,6 +1355,34 @@ void FlutterWebRTC::HandleMethodCall(
       result->NotImplemented();
     }
   }
+}
+
+void FlutterWebRTC::initLoggerCallback(RTCLoggingSeverity severity) {
+  if(eventChannelProxy == nullptr) {
+    eventChannelProxy = event_channel();
+  }
+
+  libwebrtc::LibWebRTCLogging::setLogSink(severity, [](const string& message){
+    EncodableMap info;
+    info[EncodableValue("event")] = "onLogData";
+    info[EncodableValue("data")] = message.c_string();
+    eventChannelProxy->Success(EncodableValue(info), false);
+  });
+}
+
+RTCLoggingSeverity FlutterWebRTC::str2LogSeverity(std::string str) {
+  if(str == "verbose")
+    return Verbose;
+  else if(str == "info")
+    return Info;
+  else if(str == "warning")
+    return Warning;
+  else if(str == "error")
+    return Error;
+  else if(str == "none")
+    return None;
+
+  return None;
 }
 
 }  // namespace flutter_webrtc_plus_plugin