+++ common/darwin/Classes/FlutterRTCFrameCryptor.m
@@ -38,7 +38,7 @@
 
 @implementation FlutterWebRTCPlugin (FrameCryptor)
 
-- (void)handleFrameCryptorMethodCall:(nonnull FlutterMethodCall*)call
+- (BOOL)handleFrameCryptorMethodCall:(nonnull FlutterMethodCall*)call
                               result:(nonnull FlutterResult)result {
   NSDictionary* constraints = call.arguments;
   NSString* method = call.method;
@@ -73,8 +73,10 @@
   } else if ([method isEqualToString:@"keyProviderDispose"]) {
     [self keyProviderDispose:constraints result:result];
   } else {
-    result(FlutterMethodNotImplemented);
+    return NO;
   }
+
+  return YES;
 }
 
 - (RTCCryptorAlgorithm)getAlgorithm:(NSNumber*)algorithm {
@@ -565,21 +567,21 @@
   result(@{@"result" : @"success"});
 }
 
-- (NSString*)stringFromState:(FrameCryptionState)state {
+- (NSString*)stringFromState:(RTCFrameCryptorState)state {
   switch (state) {
-    case FrameCryptionStateNew:
+    case RTCFrameCryptorStateNew:
       return @"new";
-    case FrameCryptionStateOk:
+    case RTCFrameCryptorStateOk:
       return @"ok";
-    case FrameCryptionStateEncryptionFailed:
+    case RTCFrameCryptorStateEncryptionFailed:
       return @"encryptionFailed";
-    case FrameCryptionStateDecryptionFailed:
+    case RTCFrameCryptorStateDecryptionFailed:
       return @"decryptionFailed";
-    case FrameCryptionStateMissingKey:
+    case RTCFrameCryptorStateMissingKey:
       return @"missingKey";
-    case FrameCryptionStateKeyRatcheted:
+    case RTCFrameCryptorStateKeyRatcheted:
       return @"keyRatcheted";
-    case FrameCryptionStateInternalError:
+    case RTCFrameCryptorStateInternalError:
       return @"internalError";
     default:
       return @"unknown";
@@ -590,7 +592,7 @@
 
 - (void)frameCryptor:(RTC_OBJC_TYPE(RTCFrameCryptor) *)frameCryptor
     didStateChangeWithParticipantId:(NSString*)participantId
-                          withState:(FrameCryptionState)stateChanged {
+                          withState:(RTCFrameCryptorState)stateChanged {
   if (frameCryptor.eventSink) {
     postEvent(frameCryptor.eventSink, @{
       @"event" : @"frameCryptionStateChanged",