+++ elinux/flutter_webrtc_plus/flutter_web_r_t_c_plugin.h
@@ -17,6 +17,8 @@ extern "C" {
 FLUTTER_PLUGIN_EXPORT void FlutterWebRTCPluginRegisterWithRegistrar(
     FlutterDesktopPluginRegistrarRef registrar);
 
+FLUTTER_PLUGIN_EXPORT flutter_webrtc_plugin::FlutterWebRTC* FlutterWebRTCPluginSharedInstance();
+
 #if defined(__cplusplus)
 }  // extern "C"
 #endif