+++ example/lib/main.dart
@@ -1,12 +1,11 @@
 import 'dart:core';
 
-import 'package:flutter/foundation.dart'
-    show debugDefaultTargetPlatformOverride;
 import 'package:flutter/material.dart';
 import 'package:flutter_background/flutter_background.dart';
 import 'package:flutter_webrtc_example/src/capture_frame_sample.dart';
-import 'package:flutter_webrtc_plus/flutter_webrtc_plus.dart';
 
+import 'src/adm_sample.dart';
+import 'src/data_packet_cryptor_sample.dart';
 import 'src/device_enumeration_sample.dart';
 import 'src/get_display_media_sample.dart';
 import 'src/get_user_media_sample.dart'
@@ -17,11 +16,6 @@ import 'src/route_item.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
-  if (WebRTC.platformIsDesktop) {
-    debugDefaultTargetPlatformOverride = TargetPlatform.fuchsia;
-  } else if (WebRTC.platformIsAndroid) {
-    //startForegroundService();
-  }
   runApp(MyApp());
 }
 
@@ -134,6 +128,22 @@ class _MyAppState extends State<MyApp> {
                 context,
                 MaterialPageRoute(
                     builder: (BuildContext context) => CaptureFrameSample()));
+          }),
+      RouteItem(
+          title: 'ADM Sample',
+          push: (BuildContext context) {
+            Navigator.push(
+                context,
+                MaterialPageRoute(
+                    builder: (BuildContext context) => AdmSample()));
+          }),
+      RouteItem(
+          title: 'Data Packet Cryptor Sample',
+          push: (BuildContext context) {
+            Navigator.push(
+                context,
+                MaterialPageRoute(
+                    builder: (BuildContext context) => DataPacketCryptorSample()));
           }),
     ];
   }