+++ lib/src/native/factory_impl.dart
@@ -2,10 +2,11 @@
 import 'dart:async';
 
 // Package imports:
-import 'package:webrtc_interface_plus/webrtc_interface_plus.dart';
+import 'package:webrtc_interface/webrtc_interface.dart';
 
 // Project imports:
 import '../desktop_capturer.dart';
+import 'data_packet_cryptor_impl.dart';
 import 'desktop_capturer_impl.dart';
 import 'frame_cryptor_impl.dart';
 import 'media_recorder_impl.dart';
@@ -123,3 +124,6 @@ DesktopCapturer get desktopCapturer => DesktopCapturerNative.instance;
 MediaDevices get mediaDevices => MediaDeviceNative.instance;
 
 FrameCryptorFactory get frameCryptorFactory => FrameCryptorFactoryImpl.instance;
+
+DataPacketCryptorFactory get dataPacketCryptorFactory =>
+    DataPacketCryptorFactoryImpl.instance;