+++ lib/src/native/frame_cryptor_impl.dart
@@ -6,7 +6,7 @@ import 'dart:core';
 import 'package:flutter/services.dart';
 
 // Package imports:
-import 'package:webrtc_interface_plus/webrtc_interface_plus.dart';
+import 'package:webrtc_interface/webrtc_interface.dart';
 
 // Project imports:
 import 'rtc_rtp_receiver_impl.dart';