+++ lib/src/native/media_stream_track_impl.dart
@@ -5,7 +5,7 @@ import 'dart:typed_data';
 
 // Package imports:
 import 'package:path_provider/path_provider.dart';
-import 'package:webrtc_interface_plus/webrtc_interface_plus.dart';
+import 'package:webrtc_interface/webrtc_interface.dart';
 
 // Project imports:
 import '../helper.dart';