+++ lib/src/native/navigator_impl.dart
@@ -1,5 +1,5 @@
 // Package imports:
-import 'package:webrtc_interface_plus/webrtc_interface_plus.dart';
+import 'package:webrtc_interface/webrtc_interface.dart';
 
 // Project imports:
 import 'mediadevices_impl.dart';