+++ lib/src/native/rtc_video_platform_view_controller.dart
@@ -6,14 +6,15 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
 // Package imports:
-import 'package:webrtc_interface_plus/webrtc_interface_plus.dart';
+import 'package:webrtc_interface/webrtc_interface.dart';
 
 // Project imports:
 import '../helper.dart';
+import '../video_renderer_extension.dart' show AudioControl;
 import 'utils.dart';
 
 class RTCVideoPlatformViewController extends ValueNotifier<RTCVideoValue>
-    implements VideoRenderer {
+    implements VideoRenderer, AudioControl {
   RTCVideoPlatformViewController(int viewId) : super(RTCVideoValue.empty) {
     _viewId = viewId;
   }
@@ -183,5 +184,19 @@ class RTCVideoPlatformViewController extends ValueNotifier<RTCVideoValue>
       return false;
     }
     return true;
+  }
+
+  @override
+  Future<void> setVolume(double value) async {
+    try {
+      if (_srcObject == null) {
+        throw Exception('Can\'t set volume: The MediaStream is null');
+      }
+      for (MediaStreamTrack track in _srcObject!.getAudioTracks()) {
+        await Helper.setVolume(value, track);
+      }
+    } catch (e) {
+      print('Helper.setVolume ${e.toString()}');
+    }
   }
 }