+++ lib/src/web/factory_impl.dart
@@ -1,7 +1,7 @@
 // Project imports:
 import '../desktop_capturer.dart';
 
-export 'package:dart_webrtc_plus/dart_webrtc_plus.dart'
+export 'package:dart_webrtc/dart_webrtc.dart'
     hide videoRenderer, MediaDevices, MediaRecorder;
 
 DesktopCapturer get desktopCapturer => throw UnimplementedError();