+++ CHANGELOG.md
@@ -1,6 +1,38 @@
 
 # Changelog
 
+[1.2.0+1] - 2025-09-22
+
+* Update .
+
+[1.2.0] - 2025-09-13
+
+* [Android] upgrade compileSdk to 36 as standard for 16kb pages support (#1925)
+* [Apple/Android] Local recording API for Darwin and Android (#1880)
+* [Apple/Android] Data Packet Cryptor Support.
+
+[1.1.0] - 2025-08-20
+
+* [Apple/Android] feat: Add H265/HEVC support.
+* [Mobile/Desktop] feat: Support write logs with Logger (logger package) (#1891)
+* [Android] fix: Reduce Recording Stop Delay and Prevent Encoder OOM Crashes (Android) (#1912)
+* [Native/Web] feat: small setVolume addition (#1904)
+* [Web] feat: Add texture-based video rendering for web (#1911)
+* [Android] fix: RECORDINGS - Add fallback resolutions for unsupported stream frame sizes on low-end Android devices (#1900)
+* [Android] fix: Update proguard-rules.pro (#1902)
+
+[1.0.0] - 2025-07-25
+
+* Bump version to 1.0.0
+* [Native] feat: Upgrade libwebrtc to m137. (#1877).
+* [Doc] fix: typo in package description (#1895)
+* [Android] fix: Video recording crashing and freezing on Android 14 Devices (#1886)
+* [Android] fix: Add audio recording for Android Platform (#1884)
+* [Dart] fix: Removed outdated code to avoid UI not being displayed in Windows release mode (#1890)
+* [Apple] fix: Fix compile warnings (#1887)
+* [Apple] feat: Update to m137 with audio engine (#1875)
+* [Android] fix: Ensure both video and audio tracks are added before starting the muxer (#1879)
+
 [0.14.1+5] - 2025-07-07
 
 * [Darwin] vendor: upgrade gpupixel version
@@ -22,6 +54,9 @@
 [0.14.1+1] - 2025-06-05
 
 * [Android/iOS/macOS] update latest gpupixel version
+* [Windows/Linux] feat: Add audio processing and sink API for cpp. (#1867)
+* [Linux] fix: Fixed audio device selection error for Linux. (#1864)
+* [Android] fix: Fix screen capture orientation for landscape-native devices (#1854)
 
 [0.14.1] - 2025-05-22